// src/modules/notifications/enums/notification-type.enum.ts
export enum NotificationType {
  INVITATION_REJECTED_FROM_CONTRACTOR = 'invitation_rejected_from_contractor',
  INVITATION_ACCEPTED_FROM_CONTRACTOR = 'invitation_accepted_from_contractor',
  INVITATION_REJECTED_FROM_COORDINATOR = 'invitation_rejected_from_coordinator',
  INVITATION_ACCEPTED_FROM_COORDINATOR = 'invitation_accepted_from_coordinator',
  INVITATION_RECEIVED_FROM_COORDINATOR = 'invitation_received_from_coordinator',
  CONTRACTOR_HIRED_BY_COORDINATOR = 'contractor_hired_by_coordinator',
  HIRING_CANCELLED_BY_COORDINATOR = 'hiring_cancelled_by_coordinator',
  INVITATION_CANCELLED_BY_COORDINATOR = 'invitation_cancelled_by_coordinator',
  ONE_DAY_LEFT_FOR_JOB = 'one_day_left_for_job',
}